example_bot/requirements.txt
discord.py>=1.6.0
DiscordPyFlare>=1.0.0